Rename mulshift128.h -> d2s_intrinsics.h
The idea is to move some of the basic building blocks of the Ryu
algorithm into this file which might then be optimized for different
compilers/architectures without decreasing code clarity.
