+++ ryu/BUILD
@@ -7,8 +7,8 @@ cc_library(
     "d2s.c",
     "d2s.h",
     "d2s_full_table.h",
+    "d2s_intrinsics.h",
     "digit_table.h",
-    "mulshift128.h",
     "common.h",
   ],
   hdrs = ["ryu.h"],
