+++ ryu/d2s_intrinsics.h
